Move normalization to separate files
