+++ envvar.go
@@ -21,13 +21,6 @@ const (
 	PanicOnError
 )
 
-// NormalizedName is an environment variable name that has been normalized according to rules
-// for the EnvVarSet (e.g. making variable names upper case, adding prefixes, etc).
-type NormalizedName string
-
-// NormalizeFunc normalizes an environment variable name to the form it can be found in the environment.
-type NormalizeFunc func(s *EnvVarSet, name string) NormalizedName
-
 // EnvVarSet is a set of defined environment variables.
 type EnvVarSet struct {
 	parsed            bool
@@ -104,44 +97,6 @@ func (s *EnvVarSet) out() io.Writer {
 // If output is nil, os.Stderr is used.
 func (s *EnvVarSet) SetOutput(output io.Writer) {
 	s.output = output
-}
-
-var defaultReplacer = strings.NewReplacer("-", "_", ".", "_")
-
-// DefaultNormalizeFunc formats variable names to upper case and replaces "-." chars with "_".
-func DefaultNormalizeFunc(_ *EnvVarSet, name string) NormalizedName {
-	return NormalizedName(defaultReplacer.Replace(strings.ToUpper(name)))
-}
-
-// GetNormalizeFunc returns the previously set NormalizeFunc.
-// If not set, it returns the default normalization function.
-func (s *EnvVarSet) GetNormalizeFunc() NormalizeFunc {
-	if s.normalizeNameFunc == nil {
-		return DefaultNormalizeFunc
-	}
-
-	return s.normalizeNameFunc
-}
-
-func (s *EnvVarSet) normalizeVarName(name string) NormalizedName {
-	fn := s.GetNormalizeFunc()
-
-	return fn(s, name)
-}
-
-// SetNormalizeFunc allows you to add a function which can translate variable names.
-// Environment variables added to the EnvVarSet will be translated, incoming environment variables
-// will be matched against these translated names.
-func (s *EnvVarSet) SetNormalizeFunc(fn NormalizeFunc) {
-	s.normalizeNameFunc = fn
-
-	for name, v := range s.vars {
-		normalizedName := s.normalizeVarName(v.Name)
-
-		delete(s.vars, name)
-
-		s.vars[normalizedName] = v
-	}
 }
 
 // Parse parses environment variables according to the definitions in the EnvVarSet.
